% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{FID}
\alias{FID}
\title{Fiducial summaries of simulations}
\format{
A \code{data.table} with 1500 rows and 5 columns.
}
\usage{
FID
}
\description{
A \code{data.table} containing the fiducial results of the 
  simulations whose code is given in the vignette.
}
\keyword{datasets}
