% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gfiQuantile}
\alias{gfiQuantile}
\title{Title}
\usage{
gfiQuantile(parameter, fidsamples, probs)
}
\arguments{
\item{parameter}{xx}

\item{fidsamples}{xx}

\item{probs}{xx}
}
\value{
xx
}
\description{
Title
}
\examples{
xx
}
