% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gfiSummary}
\alias{gfiSummary}
\title{Title}
\usage{
gfiSummary(fidsamples, conf = 0.95)
}
\arguments{
\item{fidsamples}{xx}

\item{conf}{confidence level}
}
\value{
xx
}
\description{
Title
}
\examples{
xx
}
