% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{MAXLHD}
\alias{MAXLHD}
\title{Maximum likelihood estimates of simulations}
\format{
A \code{matrix} with 500 rows and 3 columns.
}
\usage{
MAXLHD
}
\description{
A matrix containing the maximum likelihood estimates of the 
  simulations whose code is given in the vignette.
}
\keyword{datasets}
